requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
textstat>=0.7.3
playwright>=1.40.0
fasttext-wheel>=0.9.2
